+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
-- Add initial implementation of `artifact_storage`
+- Add gzip compression algorithm (gorarakelyan)
+- Add appending mode to the record writer (gorarakelyan)
+- Add main storage to handle artifacts — `artifact_storage` (gorarakelyan)
+- Add initial implementation of `record_storage` (narekgharibyan)
 - Update data and offset files format docs (gorarakelyan)
 - Add docs describing record file format (gorarakelyan)
 - Add `docs` directory to keep format specifications (gorarakelyan)